Add GradleRIO and IntelliJ files, move source code into new directory
