+++ .gitignore
@@ -1,41 +1,46 @@
-# WPILib #
-*/bin/
-*/build/
-*/lib/
+# WPILib
+bin/
+build/
+dist/
 sysProps.xml
+workbench.xmi
 
-# Eclipse IDE #
-*.metadata
-*/tmp/
-*.tmp
-*.bak
-*.swp
-*~.nib
-local.properties
-*.settings/
-*.loadpath
-*.project
-*.buildpath
-*.classpath
+# IDE
+*.iws
+out/
+tmp/
+.settings/
+.loadpath
+.project
+.classpath
 
-# Java specific #
+# OS Files
+.DS_Store
+.DS_Store?
+._*
+.Spotlight-V100
+.AppleDouble
+.LSOverride
+.Trashes
+ehthumbs.db
+Thumbs.db
+
+# Java
 *.class
+
+# Mobile Tools for Java (J2ME)
+.mtj.tmp/
 
 # Package Files #
 *.jar
+!gradle/*/*.jar
 *.war
 *.ear
 
-# OS generated files
-.DS_Store
-.Trashes
-._*
-Thumbs.db
-
-# Mobile Tools for Java (J2ME) #
-.mtj.tmp/
-
-# virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml #
+# virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
 
-/bin/
+# Toast/GradleRIO
+run/
+.gradle/
+gradle/.launch/
