+++ README.md
@@ -1,6 +1,12 @@
-[![Stories in Ready](https://badge.waffle.io/frc5687/Outlier1.png?label=ready&title=Ready)](https://waffle.io/frc5687/Outlier1)
+[![Build Status](https://travis-ci.org/frc5687/Outlier1.svg)](https://travis-ci.org/frc5687/Outlier1) [![Stories in Ready](https://badge.waffle.io/frc5687/Outlier1.png?label=ready&title=Ready)](https://waffle.io/frc5687/Outlier1)
 Outlier1
 =========
-[![Build Status](https://travis-ci.org/frc5687/Outlier1.svg)](https://travis-ci.org/frc5687/Outlier1)
 
 Code for FRC5687's 2015 robot
+
+## Getting Started
+Install IntelliJ IDEA Community Edition
+Clone this repository using Git
+Follow the directions for running Gradle and setting up WPILib in GradleRIO4Dummies.txt
+
+
