--- build.properties
@@ -1,4 +1,0 @@
-# Project specific information
-package=org.usfirst.frc.team5687.robot
-robot.class=${package}.Robot
-simulation.world.file=/usr/share/frcsim/worlds/GearsBotDemo.world
