+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Fri Jan 16 23:27:58 WST 2015
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-2.2.1-bin.zip
