+++ src/main/java/org/usfirst/frc/team5687/robot/commands/MoveStackerManually.java
